+++ MagFlow.DAL/Repositories/CoreScope/BaseCoreRepository.cs
@@ -2,6 +2,7 @@
 using MagFlow.EF;
 using MagFlow.EF.MultiTenancy;
 using MagFlow.Shared.Models;
+using MagFlow.Shared.Models.Interfaces;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using System;
@@ -10,6 +11,7 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Text;
 using System.Threading.Tasks;
+using static Microsoft.EntityFrameworkCore.DbLoggerCategory;
 
 namespace MagFlow.DAL.Repositories.CoreScope
 {
@@ -138,7 +140,10 @@ namespace MagFlow.DAL.Repositories.CoreScope
             {
                 using (var context = _coreContextFactory.CreateDbContext())
                 {
-                    return context.Set<TEntity>().Any(predicate);
+                    var query = context.Set<TEntity>().AsQueryable();
+                    if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                        query = query.Where(e => ((ISoftDeletable)e).RemovedAt == null);
+                    return query.Any(predicate);
                 }
             }
             catch (Exception ex)
@@ -154,7 +159,10 @@ namespace MagFlow.DAL.Repositories.CoreScope
             {
                 using (var context = _coreContextFactory.CreateDbContext())
                 {
-                    return await context.Set<TEntity>().AnyAsync(predicate);
+                    var query = context.Set<TEntity>().AsQueryable();
+                    if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                        query = query.Where(e => ((ISoftDeletable)e).RemovedAt == null);
+                    return await query.AnyAsync(predicate);
                 }
             }
             catch (Exception ex)
@@ -170,7 +178,10 @@ namespace MagFlow.DAL.Repositories.CoreScope
             {
                 using (var context = _coreContextFactory.CreateDbContext())
                 {
-                    return context.Set<TEntity>().Count(predicate);
+                    var query = context.Set<TEntity>().AsQueryable();
+                    if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                        query = query.Where(e => ((ISoftDeletable)e).RemovedAt == null);
+                    return query.Count(predicate);
                 }
             }
             catch (Exception ex)
@@ -186,7 +197,10 @@ namespace MagFlow.DAL.Repositories.CoreScope
             {
                 using (var context = _coreContextFactory.CreateDbContext())
                 {
-                    return await context.Set<TEntity>().CountAsync(predicate);
+                    var query = context.Set<TEntity>().AsQueryable();
+                    if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                        query = query.Where(e => ((ISoftDeletable)e).RemovedAt == null);
+                    return await query.CountAsync(predicate);
                 }
             }
             catch (Exception ex)
@@ -204,13 +218,35 @@ namespace MagFlow.DAL.Repositories.CoreScope
                 {
                     using (context = _coreContextFactory.CreateDbContext())
                     {
-                        context.Set<TEntity>().Remove(entity);
+                        if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                        {
+                            ((ISoftDeletable)entity).RemovedAt = DateTime.UtcNow;
+                            var isActiveProperty = entity.GetType().GetProperty(MagFlow.Shared.Constants.DatabaseConstants.ISACTIVE_PROPERTY);
+                            if (isActiveProperty != null)
+                                isActiveProperty.SetValue(entity, false);
+                            context.Set<TEntity>().Update(entity);
+                        }
+                        else
+                        {
+                            context.Set<TEntity>().Remove(entity);
+                        }
                         context.SaveChanges();
                     }
                 }
                 else
                 {
-                    context.Set<TEntity>().Remove(entity);
+                    if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                    {
+                        ((ISoftDeletable)entity).RemovedAt = DateTime.UtcNow;
+                        var isActiveProperty = entity.GetType().GetProperty(MagFlow.Shared.Constants.DatabaseConstants.ISACTIVE_PROPERTY);
+                        if (isActiveProperty != null)
+                            isActiveProperty.SetValue(entity, false);
+                        context.Set<TEntity>().Update(entity);
+                    }
+                    else
+                    {
+                        context.Set<TEntity>().Remove(entity);
+                    }
                     context.SaveChanges();
                 }
                 return Enums.Result.Success;
@@ -230,13 +266,35 @@ namespace MagFlow.DAL.Repositories.CoreScope
                 {
                     using (context = _coreContextFactory.CreateDbContext())
                     {
-                        context.Set<TEntity>().Remove(entity);
+                        if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                        {
+                            ((ISoftDeletable)entity).RemovedAt = DateTime.UtcNow;
+                            var isActiveProperty = entity.GetType().GetProperty(MagFlow.Shared.Constants.DatabaseConstants.ISACTIVE_PROPERTY);
+                            if (isActiveProperty != null)
+                                isActiveProperty.SetValue(entity, false);
+                            context.Set<TEntity>().Update(entity);
+                        }
+                        else
+                        {
+                            context.Set<TEntity>().Remove(entity);
+                        }
                         await context.SaveChangesAsync();
                     }
                 }
                 else
                 {
-                    context.Set<TEntity>().Remove(entity);
+                    if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                    {
+                        ((ISoftDeletable)entity).RemovedAt = DateTime.UtcNow;
+                        var isActiveProperty = entity.GetType().GetProperty(MagFlow.Shared.Constants.DatabaseConstants.ISACTIVE_PROPERTY);
+                        if (isActiveProperty != null)
+                            isActiveProperty.SetValue(entity, false);
+                        context.Set<TEntity>().Update(entity);
+                    }
+                    else
+                    {
+                        context.Set<TEntity>().Remove(entity);
+                    }
                     await context.SaveChangesAsync();
                 }
                 return Enums.Result.Success;
@@ -259,7 +317,21 @@ namespace MagFlow.DAL.Repositories.CoreScope
                         var entities = Find(predicate);
                         if (entities == null)
                             return Enums.Result.Error;
-                        context.Set<TEntity>().RemoveRange(entities);
+                        if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                        {
+                            foreach (var entity in entities)
+                            {
+                                ((ISoftDeletable)entity).RemovedAt = DateTime.UtcNow;
+                                var isActiveProperty = entity.GetType().GetProperty(MagFlow.Shared.Constants.DatabaseConstants.ISACTIVE_PROPERTY);
+                                if (isActiveProperty != null)
+                                    isActiveProperty.SetValue(entity, false);
+                            }
+                            context.Set<TEntity>().UpdateRange(entities);
+                        }
+                        else
+                        {
+                            context.Set<TEntity>().RemoveRange(entities);
+                        }
                         context.SaveChanges();
                     }
                 }
@@ -268,7 +340,21 @@ namespace MagFlow.DAL.Repositories.CoreScope
                     var entities = Find(predicate);
                     if (entities == null)
                         return Enums.Result.Error;
-                    context.Set<TEntity>().RemoveRange(entities);
+                    if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                    {
+                        foreach (var entity in entities)
+                        {
+                            ((ISoftDeletable)entity).RemovedAt = DateTime.UtcNow;
+                            var isActiveProperty = entity.GetType().GetProperty(MagFlow.Shared.Constants.DatabaseConstants.ISACTIVE_PROPERTY);
+                            if (isActiveProperty != null)
+                                isActiveProperty.SetValue(entity, false);
+                        }
+                        context.Set<TEntity>().UpdateRange(entities);
+                    }
+                    else
+                    {
+                        context.Set<TEntity>().RemoveRange(entities);
+                    }
                     context.SaveChanges();
                 }
                 return Enums.Result.Success;
@@ -291,7 +377,21 @@ namespace MagFlow.DAL.Repositories.CoreScope
                         var entities = Find(predicate);
                         if (entities == null)
                             return Enums.Result.Error;
-                        context.Set<TEntity>().RemoveRange(entities);
+                        if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                        {
+                            foreach (var entity in entities)
+                            {
+                                ((ISoftDeletable)entity).RemovedAt = DateTime.UtcNow;
+                                var isActiveProperty = entity.GetType().GetProperty(MagFlow.Shared.Constants.DatabaseConstants.ISACTIVE_PROPERTY);
+                                if (isActiveProperty != null)
+                                    isActiveProperty.SetValue(entity, false);
+                            }
+                            context.Set<TEntity>().UpdateRange(entities);
+                        }
+                        else
+                        {
+                            context.Set<TEntity>().RemoveRange(entities);
+                        }
                         await context.SaveChangesAsync();
                     }
                 }
@@ -300,7 +400,21 @@ namespace MagFlow.DAL.Repositories.CoreScope
                     var entities = Find(predicate);
                     if (entities == null)
                         return Enums.Result.Error;
-                    context.Set<TEntity>().RemoveRange(entities);
+                    if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                    {
+                        foreach (var entity in entities)
+                        {
+                            ((ISoftDeletable)entity).RemovedAt = DateTime.UtcNow;
+                            var isActiveProperty = entity.GetType().GetProperty(MagFlow.Shared.Constants.DatabaseConstants.ISACTIVE_PROPERTY);
+                            if (isActiveProperty != null)
+                                isActiveProperty.SetValue(entity, false);
+                        }
+                        context.Set<TEntity>().UpdateRange(entities);
+                    }
+                    else
+                    {
+                        context.Set<TEntity>().RemoveRange(entities);
+                    }
                     await context.SaveChangesAsync();
                 }
                 return Enums.Result.Success;
@@ -319,6 +433,8 @@ namespace MagFlow.DAL.Repositories.CoreScope
                 using (var context = _coreContextFactory.CreateDbContext())
                 {
                     var query = context.Set<TEntity>().AsQueryable();
+                    if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                        query = query.Where(e => ((ISoftDeletable)e).RemovedAt == null);
                     if (include != null)
                         query = include(query);
                     return query.Where(predicate).FirstOrDefault();
@@ -339,6 +455,8 @@ namespace MagFlow.DAL.Repositories.CoreScope
                 using (var context = _coreContextFactory.CreateDbContext())
                 {
                     var query = context.Set<TEntity>().AsQueryable();
+                    if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                        query = query.Where(e => ((ISoftDeletable)e).RemovedAt == null);
                     if (include != null)
                         query = include(query);
                     return await query.Where(predicate).FirstOrDefaultAsync();
@@ -358,6 +476,8 @@ namespace MagFlow.DAL.Repositories.CoreScope
                 using (var context = _coreContextFactory.CreateDbContext())
                 {
                     var query = context.Set<TEntity>().AsQueryable();
+                    if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                        query = query.Where(e => ((ISoftDeletable)e).RemovedAt == null);
                     if (include != null)
                         query = include(query);
                     if (predicate != null)
@@ -379,6 +499,8 @@ namespace MagFlow.DAL.Repositories.CoreScope
                 using (var context = _coreContextFactory.CreateDbContext())
                 {
                     var query = context.Set<TEntity>().AsQueryable();
+                    if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                        query = query.Where(e => ((ISoftDeletable)e).RemovedAt == null);
                     if (include != null)
                         query = include(query);
                     if (predicate != null)
@@ -400,6 +522,8 @@ namespace MagFlow.DAL.Repositories.CoreScope
                 using (var context = _coreContextFactory.CreateDbContext())
                 {
                     var query = context.Set<TEntity>().AsQueryable();
+                    if (typeof(ISoftDeletable).IsAssignableFrom(typeof(TEntity)))
+                        query = query.Where(e => ((ISoftDeletable)e).RemovedAt == null);
                     if (include != null)
                         query = include(query);
                     query = query.ApplyColumnFilters(options.Filters);
